Break everything

+++ src/controller.cpp
@@ -18,7 +18,7 @@ PlayerControllerState PlayerControllerState::Query(bool processMouse, bool proce
     // Freecam events
     if (freeCamera) {
         if (processMouse) {
-            input.cameraSpeedDelta = GetMouseWheelMove();
+            input.cameraSpeedDelta = (int)GetMouseWheelMove();
         }
         if (processKeyboard) {
             input.cameraNorth = IsKeyDown(KEY_W) && !IsKeyDown(KEY_S);
